refactor: move ctx from defs
